chore: potential fix for nested ifs not working
